# api/openapi.yaml
openapi: 3.0.3
info:
  title: go-httpbin
  version: 0.1.0
  description: HTTP request and response testing service with practical httpbin parity.
servers:
  - url: /
paths:
...
  /basic-auth/{user}/{password}:
    parameters:
      - $ref: "#/components/parameters/User"
      - $ref: "#/components/parameters/Password"
    get:
      operationId: BasicAuth
      responses:
        "200":
          description: Authenticated.
        "401":
          description: Unauthorized.
...
  /stream-bytes/{n}:
    parameters:
      - $ref: "#/components/parameters/N"
    get:
      operationId: StreamBytes
      responses:
        "200":
          description: Streamed random bytes.
...
  /cookies/set/{name}/{value}:
    parameters:
      - $ref: "#/components/parameters/Name"
      - $ref: "#/components/parameters/Value"
    get:
      operationId: SetCookie
      responses:
        "302":
          description: Redirect to cookie listing.
...
  /redirect/{n}:
    parameters:
      - $ref: "#/components/parameters/N"
    get:
      operationId: Redirect
      responses:
        "302":
          description: Redirect.
  /relative-redirect/{n}:
    parameters:
      - $ref: "#/components/parameters/N"
    get:
      operationId: RelativeRedirect
      responses:
        "302":
          description: Relative redirect.
  /absolute-redirect/{n}:
    parameters:
      - $ref: "#/components/parameters/N"
    get:
      operationId: AbsoluteRedirect
      responses:
        "302":
          description: Absolute redirect.
...
components:
  parameters:
...
    N:
      name: n
      in: path
      required: true
      schema:
        type: integer
    Name:
      name: name
      in: path
      required: true
      schema:
        type: string
...
    Password:
      name: password
      in: path
      required: true
      schema:
        type: string
...
    User:
      name: user
      in: path
      required: true
      schema:
        type: string
    Value:
      name: value
      in: path
      required: true
      schema:
        type: string